Fix output formatting for variables c and d in main function

+++ ex01/src/main.cpp
@@ -6,7 +6,7 @@
 /*   By: descamil <descamil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/26 17:41:26 by descamil          #+#    #+#             */
-/*   Updated: 2025/04/26 18:11:49 by descamil         ###   ########.fr       */
+/*   Updated: 2025/08/19 18:37:01 by descamil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ int main( void )
 	std::cout << "d is " << d << std::endl;
 	std::cout << "a is " << a.toInt() << " as integer" << std::endl;
 	std::cout << "b is " << b.toInt() << " as integer" << std::endl;
-	std::cout << "c is " << c.toInt() << " as integer" << std::endl;
-	std::cout << "d is " << d.toInt() << " as integer" << std::endl;
+	std::cout << "c is " << c.toFloat() << " as integer" << std::endl;
+	std::cout << "d is " << d.toFloat() << " as integer" << std::endl;
 	return 0;
 }
